+++ src/Table/components/index.js
@@ -2,3 +2,4 @@ export * from './TableToolbarContainer';
 export * from './TableTitleBar';
 export * from './TableContent';
 export * from './TableEmptyState';
+export * from './TableBulkSelectionCheckbox';